requirements.txt
#django
Django==3.1.1
django-filter==2.3.0
djangorestframework-jwt==1.11.0
djangorestframework==3.11.1
coreapi==2.3.3 # https://github.com/core-api/python-client
django-cors-headers==3.5.0 # https://pypi.org/project/django-cors-headers/
drf_yasg==1.17.1
mysqlclient==2.0.1 #sudo apt-get install libmariadbclient-dev
pyexcel_xlsx==0.5.8
django-crontab==0.7.1
xlwt==1.3.0
bs4==0.0.1
mysql-connector-python==8.0.21
django-silk==4.1.0
psycopg2-binary==2.8.6
python-decouple==3.3
reportlab==3.5.49
gunicorn==19.9.0
celery==4.4.7
redis==3.5.3
django_celery_beat==2.0.0
django_celery_results==1.2.1

firebase==3.0.1
python_jwt==3.3.0
gcloud==0.18.3
sseclient==0.0.26
pycrypto==2.6.1
requests_toolbelt==0.9.1
django-rest-framework-social-oauth2==1.1.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.9
fcm-django==0.3.4
flower==0.9.5
django-log-viewer==1.1.1
graphene~=2.1.8
graphene-django==2.11.0
user-agents==2.2.0
Pyrebase4
rest_pyotp==0.5.6